Lesson8: refactored packages
