+++ MyNotebook/app/src/main/java/com/example/mynotebook/ui/NoteDetailsFragment.java
@@ -1,4 +1,4 @@
-package com.example.mynotebook;
+package com.example.mynotebook.ui;
 
 import android.os.Bundle;
 
@@ -9,9 +9,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.example.mynotebook.data.MyNote;
+import com.example.mynotebook.R;
 import com.example.mynotebook.utils.DateUtils;
 
-import static com.example.mynotebook.Constants.CURRENT_NOTE;
+import static com.example.mynotebook.data.Constants.CURRENT_NOTE;
 
 /**
  * A simple {@link Fragment} subclass.
