+++ MyNotebook/app/src/main/res/layout/fragment_add_note.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".AddNoteFragment">
+    tools:context=".ui.AddNoteFragment">
 
     <TextView
         android:id="@+id/edit_note"
